# docs/swagger.yaml
basePath: /
definitions:
...
  learninggrouphandler.CreateLGroupResponse:
    properties:
      error:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
  learninggrouphandler.CreateLearningGroupRequest:
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  learninggrouphandler.DeleteLGroupResponse:
    properties:
      error:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
  learninggrouphandler.GetLearningGroupsResponse:
    properties:
      error:
        type: string
      learningGroups:
        $ref: '#/definitions/ssomodels.GetLGroupsResp'
      status:
        type: string
    type: object
  learninggrouphandler.GetLgByIDResponse:
    properties:
      error:
        type: string
      learningGroup:
        $ref: '#/definitions/ssomodels.GetLgByIDResp'
      status:
        type: string
    type: object
  learninggrouphandler.UpdateLGroupResponse:
    properties:
      error:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
  learninggrouphandler.UpdateLearningGroupRequest:
    properties:
      group_admins:
        items:
          type: string
        type: array
      learners:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  ssomodels.GetLGroupsResp:
    properties:
      learningGroups:
        items:
          $ref: '#/definitions/ssomodels.LearningGroup'
        type: array
    type: object
  ssomodels.GetLgByIDResp:
    properties:
      createdBy:
        type: string
      groupAdmins:
        items:
          $ref: '#/definitions/ssomodels.GroupAdmins'
        type: array
      id:
        type: string
      learners:
        items:
          $ref: '#/definitions/ssomodels.Learner'
        type: array
      modifiedBy:
        type: string
      name:
        type: string
    type: object
  ssomodels.GroupAdmins:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  ssomodels.Learner:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  ssomodels.LearningGroup:
    properties:
      created:
        type: string
      createdBy:
        type: string
      id:
        type: string
      modifiedBy:
        type: string
      name:
        type: string
      updated:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: API Support
  description: The project is only in its initial stages.
  title: Learning Platform API
  version: 0.1.0
paths:
...
  /learning_group/{id}:
    delete:
      consumes:
      - application/json
      description: This endpoint allows learning group id and delete it.
      parameters:
      - description: ID of the channel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/learninggrouphandler.DeleteLGroupResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete learning group by id
      tags:
      - learning groups
    get:
      consumes:
      - application/json
      description: This endpoint allows learning group id and returns lg info.
      parameters:
      - description: ID of the channel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/learninggrouphandler.GetLgByIDResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get learning group by id
      tags:
      - learning groups
    patch:
      consumes:
      - application/json
      description: This endpoint allows learning group id and update it.
      parameters:
      - description: ID of the channel
        in: path
        name: id
        required: true
        type: string
      - description: Getting parameters
        in: body
        name: learninggrouphandler.UpdateLearningGroupRequest
        required: true
        schema:
          $ref: '#/definitions/learninggrouphandler.UpdateLearningGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/learninggrouphandler.UpdateLGroupResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Update learning group by id
      tags:
      - learning groups
  /learning_groups:
    get:
      consumes:
      - application/json
      description: This endpoint allows user id and returns all relevant learning
        groups.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/learninggrouphandler.GetLearningGroupsResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get learning groups relevant for user
      tags:
      - learning groups
    post:
      consumes:
      - application/json
      description: This endpoint allows learning group info and create it.
      parameters:
      - description: Creating parameters
        in: body
        name: learninggrouphandler.CreateLearningGroupRequest
        required: true
        schema:
          $ref: '#/definitions/learninggrouphandler.CreateLearningGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/learninggrouphandler.CreateLGroupResponse'
        "400":
          description: Invalid data in the request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new learning group
      tags:
      - learning groups
...
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"